Move run_queries into a separate module
